feat: upload and view logs (#116)
