+++ src/file-service/domain/core/dto/file.dto.ts
@@ -2,3 +2,9 @@ export type CreateFileDTO = Express.Multer.File;
 export type CreatedFileDTO = {
   path: string;
 };
+
+export type LogDTO = {
+  userSn: number;
+  deviceUserId: string;
+  recordTime: string;
+};
